Small refactor to improve structure
