+++ belongs_to_polymorphic.gemspec
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative 'lib/belongs_to_polymorphic/version'
 
 Gem::Specification.new do |s|
@@ -15,16 +17,16 @@ Gem::Specification.new do |s|
   s.metadata['source_code_uri'] = s.homepage
   s.metadata['rubygems_mfa_required'] = 'true'
 
-  # Specify which files should be added to the gem when it is released.
-  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
   s.files = Dir['LICENSE.txt', 'README.md', 'lib/**/*']
   s.bindir        = 'exe'
   s.executables   = s.files.grep(%r{^exe/}) { |f| File.basename(f) }
   s.require_paths = ['lib']
-  s.add_development_dependency 'activerecord', '>= 5.0.0'
-  s.add_development_dependency 'rake'
-  s.add_development_dependency 'rspec'
-  s.add_development_dependency 'simplecov', '~> 0.17.1'
 
-  s.add_runtime_dependency 'activerecord', '>= 5.2', '< 8.0'
+  s.add_dependency 'activerecord', '>= 5.2', '< 8.0'
+
+  s.add_development_dependency 'database_cleaner-active_record', '~> 1.8.0'
+  s.add_development_dependency 'rake', '~> 13.0.6'
+  s.add_development_dependency 'rspec', '~> 3.11.0'
+  s.add_development_dependency 'simplecov', '~> 0.17.1'
+  s.add_development_dependency 'sqlite3', '~> 1.4.2'
 end
