+++ lib/belongs_to_polymorphic/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module BelongsToPolymorphic
-  VERSION = '0.1.1'.freeze
+  VERSION = '0.1.2'
 end
